+++ gasoline/src/java/com/stepup/gasoline/qt/bean/GasReturnDetailBean.java
@@ -16,7 +16,7 @@ public class GasReturnDetailBean {
     private String shellName;
     private int customerId;
     private String customerName;
-    private float quantity;
+    private int quantity;
     private double price;
     private double amount;
 
@@ -28,11 +28,11 @@ public class GasReturnDetailBean {
         this.id = id;
     }
 
-    public float getQuantity() {
+    public int getQuantity() {
         return quantity;
     }
 
-    public void setQuantity(float quantity) {
+    public void setQuantity(int quantity) {
         this.quantity = quantity;
     }
